+++ ddl/column_modify_test.go
@@ -41,7 +41,6 @@ import (
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/admin"
 	"github.com/pingcap/tidb/util/mock"
-	"github.com/pingcap/tidb/util/testutil"
 	"github.com/stretchr/testify/require"
 )
 
@@ -992,7 +991,7 @@ func TestCheckColumnDefaultValue(t *testing.T) {
 
 	tk.MustExec("set sql_mode='';")
 	tk.MustExec("create table text_default_text(c1 text not null default '');")
-	tk.MustQuery(`show create table text_default_text`).Check(testutil.RowsWithSep("|",
+	tk.MustQuery(`show create table text_default_text`).Check(testkit.RowsWithSep("|",
 		"text_default_text CREATE TABLE `text_default_text` (\n"+
 			"  `c1` text NOT NULL\n"+
 			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin",
@@ -1003,7 +1002,7 @@ func TestCheckColumnDefaultValue(t *testing.T) {
 	require.Empty(t, tblInfo.Meta().Columns[0].DefaultValue)
 
 	tk.MustExec("create table text_default_blob(c1 blob not null default '');")
-	tk.MustQuery(`show create table text_default_blob`).Check(testutil.RowsWithSep("|",
+	tk.MustQuery(`show create table text_default_blob`).Check(testkit.RowsWithSep("|",
 		"text_default_blob CREATE TABLE `text_default_blob` (\n"+
 			"  `c1` blob NOT NULL\n"+
 			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin",
@@ -1014,7 +1013,7 @@ func TestCheckColumnDefaultValue(t *testing.T) {
 	require.Empty(t, tblInfo.Meta().Columns[0].DefaultValue)
 
 	tk.MustExec("create table text_default_json(c1 json not null default '');")
-	tk.MustQuery(`show create table text_default_json`).Check(testutil.RowsWithSep("|",
+	tk.MustQuery(`show create table text_default_json`).Check(testkit.RowsWithSep("|",
 		"text_default_json CREATE TABLE `text_default_json` (\n"+
 			"  `c1` json NOT NULL DEFAULT 'null'\n"+
 			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin",
