+++ util/testkit/testkit.go
@@ -32,10 +32,8 @@ import (
 	"github.com/pingcap/tidb/parser/model"
 	"github.com/pingcap/tidb/parser/terror"
 	"github.com/pingcap/tidb/session"
-	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/sqlexec"
-	"github.com/pingcap/tidb/util/testutil"
 )
 
 // TestKit is a utility to run sql test.
@@ -119,15 +117,6 @@ func NewTestKit(c *check.C, store kv.Storage) *TestKit {
 	return &TestKit{
 		c:     c,
 		store: store,
-	}
-}
-
-// NewTestKitWithSession returns a new *TestKit with a session.
-func NewTestKitWithSession(c *check.C, store kv.Storage, se session.Session) *TestKit {
-	return &TestKit{
-		c:     c,
-		store: store,
-		Se:    se,
 	}
 }
 
@@ -395,7 +384,7 @@ func (tk *TestKit) ResultSetToResultWithCtx(ctx context.Context, rs sqlexec.Reco
 
 // Rows is similar to RowsWithSep, use white space as separator string.
 func Rows(args ...string) [][]interface{} {
-	return testutil.RowsWithSep(" ", args...)
+	return RowsWithSep(" ", args...)
 }
 
 // GetTableID gets table ID by name.
@@ -407,9 +396,17 @@ func (tk *TestKit) GetTableID(tableName string) int64 {
 	return tbl.Meta().ID
 }
 
-// WithPruneMode run test case under prune mode.
-func WithPruneMode(tk *TestKit, mode variable.PartitionPruneMode, f func()) {
-	tk.MustExec("set @@tidb_partition_prune_mode=`" + string(mode) + "`")
-	tk.MustExec("set global tidb_partition_prune_mode=`" + string(mode) + "`")
-	f()
+// RowsWithSep is a convenient function to wrap args to a slice of []interface.
+// The arg represents a row, split by sep.
+func RowsWithSep(sep string, args ...string) [][]interface{} {
+	rows := make([][]interface{}, len(args))
+	for i, v := range args {
+		strs := strings.Split(v, sep)
+		row := make([]interface{}, len(strs))
+		for j, s := range strs {
+			row[j] = s
+		}
+		rows[i] = row
+	}
+	return rows
 }
